Refactored the way operators are assigned
Created temporary function (which will be renamed later) to A) focus on grabbing the second to last operator clicked and B) the last operator clicked that is NOT the = operator.

This is so when clicking an operator math is done ONLY on the second operator click and uses the second to last operator (i.e. 1+2-3.  On the  click of - then math is done using the +.  We technically do not care what the current operator clicked is i.e. the - until the next operator is clicked)

Also renamed the variable from "previousOperator" to "lastOperatorThatIsNotEqual" which is a better name.  This is primarily used for consecutive = clicks so math will keep being done using the last operator that was clicked that was NOT an =.  Need this because once you have at least two = clicks then the previous operator will be an = and math will not be correctly done