Chaining operator clicks together works correctly
Added the operate function for when an operator button is clicked so the  last clicked operator is passed to do math on the next operator click (i.e. 1+2- will do 1+2 on the click of the - btn to give a total which is then used on the next round of math.  So 1+2 = 3 and then 3 is used as part of the next equation)

Moved the assignNumbers function up so it's below the assignOperators function so the functions are grouped together in a more efficient order