Fixed bug relating to math performed after delete btn pressed
Ex: 1+2=3 (delete clicked) 3 (start of new equation)

Added if statement in delete function to reset the operator array to empty.  This way on the next operator click the operator passed is undefined, so no math is being done and only the operator clicked is being displayed.  Then on either the second operator click OR the = click, math will be done since a valid operator will then be passed at that point to complete equation