Fix bug relating to using delete btn
Scenario: 100 (delete so 10) + 200 (delete so 20) - (minus).  So this would read 10+20- but using delete on both number1 & number2 before hitting enter  was not passing the correct operator to perform the math

Updated the operatorBtnClicked function to add a nested if/else statement when the secondToLastOperator that was clicked is Delete/Backspace.  It looks to see if number2 was clicked.  If it wasn't just display the operator that  was last clicked.  If both number1 & number2 were clicked and involving deletes then pass the third operator back in the array for the math