Fixed bug relating to hitting enter after dividing by 0 & updated delete btn function
When doing 4/0 the calc displays the error but if you clicked Enter on keyboard as the next click you were getting dispaly of 0/=0.  Equal btn (=) is disabled so the issue was related to enter on keyboard.  Updated logic so if the last operator clicked was enter from keyboard, number 2 is 0 and last operator clicked is / then return and stop running the function

Updated delete btn function so when clicking it will re-enable the operator & equal btns on screen